# docs/swagger.yaml
basePath: /v1/users
definitions:
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.CreateDirectory:
    properties:
      channel_id:
        type: string
      id:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.DeleteDirectory:
    properties:
      channel_id:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.Directory:
    properties:
      channel_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.UpdateDirectory:
    properties:
      channel_id:
        type: string
      id:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is PingVi selective forwarding unit.
  termsOfService: http://swagger.io/terms/
  title: PingVi selective forwarding unit
  version: "1.0"
paths:
...
  /directorys:
    delete:
      consumes:
      - application/json
      description: Permanently deletes a directory from a channel. Requires 'manage_directories'
        permission (permission[4] == '1').
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Directory deletion data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.DeleteDirectory'
      produces:
      - application/json
      responses:
        "200":
          description: "Directory ID\"\t\t\t\texample({\"directory_id\":\"123e4567-e89b-12d3-a456-426614174000\"})"
          schema:
            additionalProperties: true
            type: object
        "400":
          description: "Invalid request\"\t\t\texample({\"error\":\"invalid body\"})"
          schema:
            additionalProperties: true
            type: object
        "403":
          description: "Insufficient permissions\"\texample({\"error\":\"not enough
            rights\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\t\texample({\"error\":\"failed to
            delete directory\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a directory
      tags:
      - messenger-directories
    post:
      consumes:
      - application/json
      description: Creates a new directory in a channel. Requires 'manage_directories'
        permission (permission[4] == '1').
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Directory creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.CreateDirectory'
      produces:
      - application/json
      responses:
        "201":
          description: Created directory
          schema:
            $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.Directory'
        "400":
          description: "Invalid request\"\t\t\texample({\"error\":\"invalid body\"})"
          schema:
            additionalProperties: true
            type: object
        "403":
          description: "Insufficient permissions\"\texample({\"error\":\"not enough
            rights\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\t\texample({\"error\":\"failed to
            create directory\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new directory in a channel
      tags:
      - messenger-directories
    put:
      consumes:
      - application/json
      description: Updates an existing directory in a channel. Requires 'manage_directories'
        permission (permission[4] == '1').
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Directory update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.UpdateDirectory'
      produces:
      - application/json
      responses:
        "200":
          description: Updated directory
          schema:
            $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.Directory'
        "400":
          description: "Invalid request\"\t\t\texample({\"error\":\"invalid body\"})"
          schema:
            additionalProperties: true
            type: object
        "403":
          description: "Insufficient permissions\"\texample({\"error\":\"not enough
            rights\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\t\texample({\"error\":\"failed to
            update directory\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a directory
      tags:
      - messenger-directories
  /directorys/{channel_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all directories from a specific channel.
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Channel ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: channel_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of directories
          schema:
            items:
              $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.Directory'
            type: array
        "400":
          description: "Invalid channel ID\"\texample({\"error\":\"invalid channel
            id\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to get
            directories\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all directories in a channel
      tags:
      - messenger-directories
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"